First pass assigning the correct pins throughout the conifg. Added comments & descriptions as required to these.
There's still some confusion as to how the '[probe]' and '[bltouch]' sections interact.
According to the Klipper reference, both of these sections are supposed to set up the 'probe:z_virtual_endstop' pin, but this wasn't done until I specified a minimal '[probe]' section in addition to a '[bltouch]' section.
Got some ideas on what to try next, but I'll be tackling that once I've established comms wwith the MCU,